# docs/protocol/local-admin-api-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Private Sync Gate Local Administration API",
    "version": "1.0.0",
    "description": "Loopback-only, owner-controlled administration API for the local daemon."
  },
  "servers": [
    { "url": "http://127.0.0.1:47820", "description": "Local daemon loopback listener" }
  ],
  "paths": {
...
    "/api/v1/status": {
      "get": {
        "operationId": "getStatus",
        "summary": "Read daemon status",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Current status", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/StatusResponse" } } } }, "401": { "$ref": "#/components/responses/Unauthorized" }, "503": { "$ref": "#/components/responses/Unavailable" } }
      }
    },
    "/api/v1/diagnostics": {
      "get": {
        "operationId": "getDiagnostics",
        "summary": "Read bounded diagnostics",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/DiagnosticsLimit" }],
        "responses": { "200": { "description": "Diagnostic report", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DiagnosticReport" } } } }, "401": { "$ref": "#/components/responses/Unauthorized" }, "503": { "$ref": "#/components/responses/Unavailable" } }
      }
    },
    "/api/v1/shares": {
      "get": {
        "operationId": "listShares",
        "summary": "List configured shares",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Limit" }, { "$ref": "#/components/parameters/Cursor" }],
        "responses": { "200": { "description": "Share page", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SharePage" } } } }, "401": { "$ref": "#/components/responses/Unauthorized" }, "503": { "$ref": "#/components/responses/Unavailable" } }
      }
    },
    "/api/v1/devices": {
      "get": {
        "operationId": "listDevices",
        "summary": "List paired devices",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Limit" }, { "$ref": "#/components/parameters/Cursor" }],
        "responses": { "200": { "description": "Device page", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DevicePage" } } } }, "401": { "$ref": "#/components/responses/Unauthorized" }, "503": { "$ref": "#/components/responses/Unavailable" } }
      }
    },
    "/api/v1/devices/{device_id}": {
      "get": {
        "operationId": "getDevice",
        "summary": "Read one paired device",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/DeviceID" }],
        "responses": { "200": { "description": "Device details", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeviceDetail" } } } }, "401": { "$ref": "#/components/responses/Unauthorized" }, "404": { "$ref": "#/components/responses/NotFound" }, "503": { "$ref": "#/components/responses/Unavailable" } }
      }
    },
    "/api/v1/federation/nodes": {
      "get": {
        "operationId": "listFederatedNodes",
        "summary": "List sanitized signed status replicas for paired nodes with active read-only grants",
        "security": [{ "bearerAuth": [] }, { "browserSession": [] }],
        "responses": { "200": { "description": "Visible paired-node status replicas", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FederatedNodePage" } } } }, "401": { "$ref": "#/components/responses/Unauthorized" }, "503": { "$ref": "#/components/responses/Unavailable" } }
      }
    },
    "/api/v1/jobs": {
      "get": {
        "operationId": "listJobs",
        "summary": "List synchronization jobs",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Limit" }, { "$ref": "#/components/parameters/Cursor" }],
        "responses": { "200": { "description": "Job page", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JobPage" } } } }, "401": { "$ref": "#/components/responses/Unauthorized" }, "503": { "$ref": "#/components/responses/Unavailable" } }
      }
    },
...
    "/api/v1/audit-events": {
      "get": {
        "operationId": "listAuditEvents",
        "summary": "List administrative audit events",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Limit" }, { "$ref": "#/components/parameters/Cursor" }],
        "responses": { "200": { "description": "Audit event page", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuditEventPage" } } } }, "401": { "$ref": "#/components/responses/Unauthorized" }, "503": { "$ref": "#/components/responses/Unavailable" } }
      }
    },
    "/api/v1/project-migrations/preflight": {
      "post": {
        "operationId": "preflightProjectMigration",
        "summary": "Read-only eligibility and confirmation check for an existing source share",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProjectMigrationRequest" } } } },
        "responses": { "200": { "description": "Migration preflight", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProjectMigrationPreflight" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "404": { "$ref": "#/components/responses/NotFound" }, "503": { "$ref": "#/components/responses/Unavailable" } }
      }
    },
    "/api/v1/project-migrations/apply": {
      "post": {
        "operationId": "applyProjectMigration",
        "summary": "Apply an unchanged confirmed project migration preflight",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProjectMigrationApplyRequest" } } } },
        "responses": { "200": { "description": "Migration applied or already applied", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProjectMigrationApplyResult" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "404": { "$ref": "#/components/responses/NotFound" }, "409": { "$ref": "#/components/responses/Conflict" }, "503": { "$ref": "#/components/responses/Unavailable" } }
      }
    },
    "/api/v1/projects": {
      "get": {
        "operationId": "listProjects",
        "summary": "List registered Agent Projects without local root paths",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Limit" }, { "$ref": "#/components/parameters/Cursor" }],
        "responses": { "200": { "description": "Project page", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProjectPage" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "503": { "$ref": "#/components/responses/Unavailable" } }
      }
    },
...
    "/api/v1/orchestration/trades": { "get": { "operationId": "listTrades", "summary": "List versioned trade inventory", "security": [{ "bearerAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/Limit" }, { "$ref": "#/components/parameters/Cursor" }], "responses": { "200": { "description": "Trade page" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "503": { "$ref": "#/components/responses/Unavailable" } } } },
    "/api/v1/orchestration/workers": { "get": { "operationId": "listWorkers", "summary": "List versioned worker inventory", "security": [{ "bearerAuth": [] }, { "browserSession": [] }], "parameters": [{ "$ref": "#/components/parameters/Limit" }, { "$ref": "#/components/parameters/Cursor" }], "responses": { "200": { "description": "Worker page" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "503": { "$ref": "#/components/responses/Unavailable" } } } },
    "/api/v1/orchestration/capabilities": { "get": { "operationId": "getSetupCapabilities", "summary": "Read configured setup capabilities; execution remains disabled", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Capability inventory" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "503": { "$ref": "#/components/responses/Unavailable" } } } },
    "/api/v1/orchestration/projects": { "get": { "operationId": "listLocalProjects", "summary": "List sanitized local project operation status", "security": [{ "bearerAuth": [] }, { "browserSession": [] }], "parameters": [{ "$ref": "#/components/parameters/Limit" }, { "$ref": "#/components/parameters/Cursor" }], "responses": { "200": { "description": "Bounded local project page", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LocalProjectPage" } } } }, "401": { "$ref": "#/components/responses/Unauthorized" }, "503": { "$ref": "#/components/responses/Unavailable" } } } },
...
    "/api/v1/orchestration/nodes": { "get": { "operationId": "listOrchestrationNodes", "summary": "List sanitized node inventory", "security": [{ "bearerAuth": [] }, { "browserSession": [] }], "parameters": [{ "$ref": "#/components/parameters/Limit" }, { "$ref": "#/components/parameters/Cursor" }], "responses": { "200": { "description": "Node page" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "503": { "$ref": "#/components/responses/Unavailable" } } } },
...
    "/api/v1/pairing/invitations": {
      "post": {
        "operationId": "createPairingInvitation",
        "summary": "Create a short-lived pairing invitation",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/InvitationRequest" } } } },
        "responses": { "201": { "description": "Invitation created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Invitation" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "503": { "$ref": "#/components/responses/Unavailable" } }
      }
    },
    "/api/v1/pairing/inspect": {
      "post": {
        "operationId": "inspectPairingInvitation",
        "summary": "Inspect an invitation without accepting it",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EncodedInvitation" } } } },
        "responses": { "200": { "description": "Invitation metadata", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/InvitationInspection" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "503": { "$ref": "#/components/responses/Unavailable" } }
      }
    },
    "/api/v1/pairing/acceptances": {
      "post": {
        "operationId": "acceptPairingInvitation",
        "summary": "Accept a pairing invitation",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AcceptanceRequest" } } } },
        "responses": { "200": { "description": "Pairing accepted or already accepted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Acceptance" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "409": { "$ref": "#/components/responses/Conflict" }, "503": { "$ref": "#/components/responses/Unavailable" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "local-admin-token" },
      "browserSession": { "type": "apiKey", "in": "cookie", "name": "syncgate_browser_session" }
    },
    "parameters": {
      "Limit": { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 50 } },
      "Cursor": { "name": "cursor", "in": "query", "schema": { "type": "string", "maxLength": 512 } },
      "DeviceID": { "name": "device_id", "in": "path", "required": true, "schema": { "type": "string", "minLength": 1, "maxLength": 128 } },
...
      "DiagnosticsLimit": { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 50 } }
    },
    "responses": {
      "BadRequest": { "description": "Invalid request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Unauthorized": { "description": "Missing or invalid local credential", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "NotFound": { "description": "Resource not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Conflict": { "description": "Command conflicts with current state", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Unavailable": { "description": "Daemon service unavailable", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    },
    "schemas": {
...
      "StatusResponse": { "type": "object", "required": ["status", "api_version", "device_id", "fingerprint", "started_at", "lifecycle", "active_share_count", "peer_executor_enabled", "queue"], "properties": { "status": { "type": "string", "enum": ["starting", "running", "draining"] }, "api_version": { "type": "string", "const": "v1" }, "device_id": { "type": "string" }, "fingerprint": { "type": "string" }, "started_at": { "type": "string", "format": "date-time" }, "lifecycle": { "type": "string", "enum": ["starting", "running", "draining"] }, "active_share_count": { "type": "integer", "minimum": 0 }, "peer_executor_enabled": { "type": "boolean" }, "queue": { "$ref": "#/components/schemas/QueueSummary" } } },
...
      "DiagnosticReport": { "type": "object", "required": ["generated_at", "recent_scans", "work", "ignored_paths"], "properties": { "generated_at": { "type": "string", "format": "date-time" }, "recent_scans": { "type": "array", "maxItems": 200, "items": { "$ref": "#/components/schemas/ScanDiagnostic" } }, "work": { "type": "array", "maxItems": 200, "items": { "$ref": "#/components/schemas/WorkDiagnostic" } }, "ignored_paths": { "type": "array", "maxItems": 200, "items": { "$ref": "#/components/schemas/IgnoreDiagnostic" } } } },
...
      "SharePage": { "type": "object", "required": ["items", "page"], "properties": { "items": { "type": "array", "maxItems": 200, "items": { "$ref": "#/components/schemas/Share" } }, "page": { "$ref": "#/components/schemas/PageInfo" } } },
      "DevicePage": { "type": "object", "required": ["items", "page"], "properties": { "items": { "type": "array", "maxItems": 200, "items": { "$ref": "#/components/schemas/Device" } }, "page": { "$ref": "#/components/schemas/PageInfo" } } },
...
      "FederatedNodePage": { "type": "object", "additionalProperties": false, "required": ["items"], "properties": { "items": { "type": "array", "maxItems": 200, "items": { "$ref": "#/components/schemas/FederatedNode" } } } },
      "JobPage": { "type": "object", "required": ["items", "page"], "properties": { "items": { "type": "array", "maxItems": 200, "items": { "$ref": "#/components/schemas/Job" } }, "page": { "$ref": "#/components/schemas/PageInfo" } } },
      "AuditEventPage": { "type": "object", "required": ["items", "page"], "properties": { "items": { "type": "array", "maxItems": 200, "items": { "$ref": "#/components/schemas/AuditEvent" } }, "page": { "$ref": "#/components/schemas/PageInfo" } } },
      "ProjectMigrationRequest": { "type": "object", "additionalProperties": false, "required": ["share_id", "project_id", "name"], "properties": { "share_id": { "type": "string", "minLength": 1, "maxLength": 128 }, "project_id": { "type": "string", "minLength": 1, "maxLength": 128 }, "name": { "type": "string", "minLength": 1, "maxLength": 256 } } },
      "ProjectMigrationApplyRequest": { "type": "object", "additionalProperties": false, "required": ["share_id", "project_id", "name", "confirmation"], "properties": { "share_id": { "type": "string", "minLength": 1, "maxLength": 128 }, "project_id": { "type": "string", "minLength": 1, "maxLength": 128 }, "name": { "type": "string", "minLength": 1, "maxLength": 256 }, "confirmation": { "type": "string", "pattern": "^[a-f0-9]{64}$" } } },
...
      "ProjectMigrationPreflight": { "type": "object", "additionalProperties": false, "required": ["status", "share_id", "project_id", "name", "root_identity", "issues", "excluded_paths", "required_ignore_patterns", "missing_configured_ignore_patterns", "configuration_change_required", "expected_portable_records", "confirmation"], "properties": { "status": { "type": "string", "enum": ["ready", "already_complete", "blocked"] }, "share_id": { "type": "string" }, "project_id": { "type": "string" }, "name": { "type": "string" }, "root_identity": { "type": "string", "pattern": "^sha256:[a-f0-9]{64}$" }, "issues": { "type": "array", "maxItems": 64, "items": { "$ref": "#/components/schemas/ProjectMigrationIssue" } }, "excluded_paths": { "type": "array", "maxItems": 64, "items": { "type": "string", "maxLength": 4096 } }, "required_ignore_patterns": { "type": "array", "maxItems": 32, "items": { "type": "string", "maxLength": 256 } }, "missing_configured_ignore_patterns": { "type": "array", "maxItems": 32, "items": { "type": "string", "maxLength": 256 } }, "configuration_change_required": { "type": "boolean" }, "expected_portable_records": { "type": "array", "maxItems": 16, "items": { "type": "string", "maxLength": 4096 } }, "existing_project_id": { "type": "string" }, "confirmation": { "type": "string", "pattern": "^[a-f0-9]{64}$" } } },
      "ProjectMigrationApplyResult": { "type": "object", "additionalProperties": false, "required": ["status", "project_id", "share_id", "created_records", "projected_events", "projected_artifacts", "insight_count", "event_watermark", "audit_event_id", "scan_requested"], "properties": { "status": { "type": "string", "enum": ["applied", "already_applied"] }, "project_id": { "type": "string" }, "share_id": { "type": "string" }, "created_records": { "type": "integer", "minimum": 0 }, "projected_events": { "type": "integer", "minimum": 0 }, "projected_artifacts": { "type": "integer", "minimum": 0 }, "insight_count": { "type": "integer", "minimum": 0 }, "event_watermark": { "type": "string" }, "audit_event_id": { "type": "string" }, "scan_requested": { "type": "boolean" } } },
...
      "ProjectPage": { "type": "object", "required": ["items", "page"], "properties": { "items": { "type": "array", "maxItems": 200, "items": { "$ref": "#/components/schemas/Project" } }, "page": { "$ref": "#/components/schemas/PageInfo" } } },
...
      "LocalProjectPage": { "type": "object", "additionalProperties": false, "required": ["items", "page"], "properties": { "items": { "type": "array", "maxItems": 200, "items": { "$ref": "#/components/schemas/LocalProject" } }, "page": { "$ref": "#/components/schemas/PageInfo" } } },
...
      "InvitationRequest": { "type": "object", "required": ["ttl_seconds"], "additionalProperties": false, "properties": { "ttl_seconds": { "type": "integer", "minimum": 1, "maximum": 86400 }, "capabilities": { "type": "array", "maxItems": 16, "items": { "type": "string", "maxLength": 64 } } } },
      "Invitation": { "type": "object", "required": ["invitation", "fingerprint", "one_time_code", "expires_at"], "additionalProperties": false, "properties": { "invitation": { "type": "string", "maxLength": 65536 }, "fingerprint": { "type": "string" }, "one_time_code": { "type": "string" }, "expires_at": { "type": "string", "format": "date-time" } } },
      "EncodedInvitation": { "type": "object", "required": ["invitation"], "additionalProperties": false, "properties": { "invitation": { "type": "string", "minLength": 1, "maxLength": 65536 } } },
...
      "Acceptance": { "type": "object", "required": ["accepted", "already_accepted", "device_id", "device_name", "fingerprint"], "additionalProperties": false, "properties": { "accepted": { "type": "boolean", "const": true }, "already_accepted": { "type": "boolean" }, "device_id": { "type": "string" }, "device_name": { "type": "string" }, "fingerprint": { "type": "string" } } },
...
    }
  }
}